ported hashes
