+++ src/hash.rs
