+++ src/hash/formats.rs
