+++ src/hash/formats/text.rs
