+++ src/hash/hash_serde.rs
