+++ src/hash/substitutions.rs
